requirements.txt
discord.py==2.2.2
googletrans==3.1.0a0
python-dotenv==1.0.0